refactor mcp4728 component
