+++ esphome/components/mcp4728/mcp4728.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include "esphome/core/component.h"
-#include "esphome/components/output/float_output.h"
 #include "esphome/components/i2c/i2c.h"
 
 namespace esphome {
@@ -62,29 +61,6 @@ class MCP4728Component : public Component, public i2c::I2CDevice {
   DACInputData reg_[4];
   bool store_in_eeprom_ = false;
   bool update_ = false;
-};
-
-class MCP4728Channel : public output::FloatOutput {
- public:
-  MCP4728Channel(MCP4728Component *parent, MCP4728ChannelIdx channel, MCP4728Vref vref, MCP4728Gain gain,
-                 MCP4728PwrDown pwrdown)
-      : parent_(parent), channel_(channel), vref_(vref), gain_(gain), pwrdown_(pwrdown) {
-    // update VREF
-    parent->select_vref_(channel, vref_);
-    // update PD
-    parent->select_power_down_(channel, pwrdown_);
-    // update GAIN
-    parent->select_gain_(channel, gain_);
-  }
-
- protected:
-  void write_state(float state) override;
-
-  MCP4728Component *parent_;
-  MCP4728ChannelIdx channel_;
-  MCP4728Vref vref_;
-  MCP4728Gain gain_;
-  MCP4728PwrDown pwrdown_;
 };
 
 }  // namespace mcp4728
