+++ esphome/components/mcp4728/output/__init__.py
@@ -2,12 +2,11 @@ import esphome.codegen as cg
 import esphome.config_validation as cv
 from esphome.components import output
 from esphome.const import CONF_CHANNEL, CONF_ID, CONF_GAIN
-from . import MCP4728Component, mcp4728_ns
+from .. import MCP4728Component, CONF_MCP4728_ID, mcp4728_ns
 
 DEPENDENCIES = ["mcp4728"]
 
 MCP4728Channel = mcp4728_ns.class_("MCP4728Channel", output.FloatOutput)
-CONF_MCP4728_ID = "mcp4728_id"
 CONF_VREF = "vref"
 CONF_POWER_DOWN = "power_down"
